requirements.txt
# requirements.txt
git+https://github.com/speedyapply/JobSpy.git
numpy==1.26.3
pandas==2.2.2
pyyaml==6.0.1
requests==2.32.3
loguru==0.7.2
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
notebook==7.2.2
ipykernel==6.29.5
matplotlib==3.8.3
seaborn==0.13.2